Resolve some of the points raised during review:
- Fix alignment of `=` in docs
- Share functions for regional logic in both cluster and node pool.
The extra functions have been moved to the regional_utils.
- Removed unused getParent
- Moved getRegion back to bottom of node_pool
- Change cluster back to required for the node pool
- Move setting node pool ID during creation back to above the wait
- Used the node pool info within the exists function for node pool
